requirements.txt
numpy==1.23.5

tensorflow-cpu==2.10

tensorflow-directml-plugin